Move geo utils files into a geoengine utiles sub-directory
This is a minor move really, but the aim is to have all geo functionality
within the geoengine directory.
