+++ app/src/main/java/org/scottishtecharmy/soundscape/SoundscapeServiceConnection.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
 import org.scottishtecharmy.soundscape.services.SoundscapeBinder
 import org.scottishtecharmy.soundscape.services.SoundscapeService
-import org.scottishtecharmy.soundscape.utils.TileGrid
+import org.scottishtecharmy.soundscape.geoengine.utils.TileGrid
 import javax.inject.Inject
 
 @ActivityRetainedScoped
