+++ app/src/main/java/org/scottishtecharmy/soundscape/geoengine/utils/BoEvent.kt
@@ -1,4 +1,4 @@
-package org.scottishtecharmy.soundscape.utils
+package org.scottishtecharmy.soundscape.geoengine.utils
 
 import java.util.ArrayList
