+++ app/src/main/java/org/scottishtecharmy/soundscape/geoengine/utils/BoPoint.kt
@@ -1,4 +1,4 @@
-package org.scottishtecharmy.soundscape.utils
+package org.scottishtecharmy.soundscape.geoengine.utils
 
 data class BoPoint(
     var x_coord: Double,
