+++ app/src/main/java/org/scottishtecharmy/soundscape/geoengine/utils/Quadrant.kt
@@ -1,4 +1,4 @@
-package org.scottishtecharmy.soundscape.utils
+package org.scottishtecharmy.soundscape.geoengine.utils
 
 class Quadrant(private val heading: Double) {
     val left: Double
