+++ app/src/test/java/org/scottishtecharmy/soundscape/GeoUtilsTest.kt
@@ -7,42 +7,37 @@ import org.scottishtecharmy.soundscape.geojsonparser.geojson.MultiPoint
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.MultiPolygon
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Point
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Polygon
-import org.scottishtecharmy.soundscape.utils.bearingFromTwoPoints
-import org.scottishtecharmy.soundscape.utils.createTriangleFOV
-import org.scottishtecharmy.soundscape.utils.distance
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxCorners
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxOfLineString
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxOfMultiLineString
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxOfMultiPoint
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxOfMultiPolygon
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxOfPoint
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxOfPolygon
-import org.scottishtecharmy.soundscape.utils.getCenterOfBoundingBox
-import org.scottishtecharmy.soundscape.utils.getDestinationCoordinate
-import org.scottishtecharmy.soundscape.utils.getPixelXY
-import org.scottishtecharmy.soundscape.utils.getPolygonOfBoundingBox
-import org.scottishtecharmy.soundscape.utils.getQuadKey
-import org.scottishtecharmy.soundscape.utils.getQuadrants
-import org.scottishtecharmy.soundscape.utils.getReferenceCoordinate
-import org.scottishtecharmy.soundscape.utils.groundResolution
-import org.scottishtecharmy.soundscape.utils.mapSize
-import org.scottishtecharmy.soundscape.utils.pixelXYToLatLon
-import org.scottishtecharmy.soundscape.utils.polygonContainsCoordinates
+import org.scottishtecharmy.soundscape.geoengine.utils.bearingFromTwoPoints
+import org.scottishtecharmy.soundscape.geoengine.utils.createTriangleFOV
+import org.scottishtecharmy.soundscape.geoengine.utils.distance
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxCorners
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxOfLineString
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxOfMultiLineString
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxOfMultiPoint
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxOfMultiPolygon
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxOfPoint
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxOfPolygon
+import org.scottishtecharmy.soundscape.geoengine.utils.getCenterOfBoundingBox
+import org.scottishtecharmy.soundscape.geoengine.utils.getDestinationCoordinate
+import org.scottishtecharmy.soundscape.geoengine.utils.getPixelXY
+import org.scottishtecharmy.soundscape.geoengine.utils.getPolygonOfBoundingBox
+import org.scottishtecharmy.soundscape.geoengine.utils.getQuadKey
+import org.scottishtecharmy.soundscape.geoengine.utils.getQuadrants
+import org.scottishtecharmy.soundscape.geoengine.utils.getReferenceCoordinate
+import org.scottishtecharmy.soundscape.geoengine.utils.groundResolution
+import org.scottishtecharmy.soundscape.geoengine.utils.mapSize
+import org.scottishtecharmy.soundscape.geoengine.utils.pixelXYToLatLon
+import org.scottishtecharmy.soundscape.geoengine.utils.polygonContainsCoordinates
 import org.junit.Assert
 import org.junit.Test
-import org.scottishtecharmy.soundscape.utils.BentleyOttmann
-import org.scottishtecharmy.soundscape.utils.BoPoint
-import org.scottishtecharmy.soundscape.utils.BoSegment
-import org.scottishtecharmy.soundscape.utils.calculateCenterOfCircle
-import org.scottishtecharmy.soundscape.utils.distanceToPolygon
-import org.scottishtecharmy.soundscape.utils.getXYTile
-import org.scottishtecharmy.soundscape.utils.isBetween
-import org.scottishtecharmy.soundscape.utils.lineStringsIntersect
-import org.scottishtecharmy.soundscape.utils.straightLinesIntersect
-import org.scottishtecharmy.soundscape.utils.straightLinesIntersectLngLatAlt
-import kotlin.math.abs
-import kotlin.math.min
-import kotlin.math.sqrt
+import org.scottishtecharmy.soundscape.geoengine.utils.BentleyOttmann
+import org.scottishtecharmy.soundscape.geoengine.utils.BoPoint
+import org.scottishtecharmy.soundscape.geoengine.utils.BoSegment
+import org.scottishtecharmy.soundscape.geoengine.utils.calculateCenterOfCircle
+import org.scottishtecharmy.soundscape.geoengine.utils.distanceToPolygon
+import org.scottishtecharmy.soundscape.geoengine.utils.lineStringsIntersect
+import org.scottishtecharmy.soundscape.geoengine.utils.straightLinesIntersect
+import org.scottishtecharmy.soundscape.geoengine.utils.straightLinesIntersectLngLatAlt
 
 
 class GeoUtilsTest {
