+++ app/src/test/java/org/scottishtecharmy/soundscape/KalmanFilterTest.kt
@@ -4,8 +4,8 @@ import org.junit.Test
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
 import org.scottishtecharmy.soundscape.kalman.KalmanFilter
-import org.scottishtecharmy.soundscape.utils.distance
-import org.scottishtecharmy.soundscape.utils.getDestinationCoordinate
+import org.scottishtecharmy.soundscape.geoengine.utils.distance
+import org.scottishtecharmy.soundscape.geoengine.utils.getDestinationCoordinate
 
 class KalmanFilterTest {
