+++ app/src/test/java/org/scottishtecharmy/soundscape/RoundaboutsTest.kt
@@ -12,35 +12,32 @@ import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.MultiLineString
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Point
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Polygon
-import org.scottishtecharmy.soundscape.utils.RelativeDirections
-import org.scottishtecharmy.soundscape.utils.calculateCenterOfCircle
-import org.scottishtecharmy.soundscape.utils.createTriangleFOV
-import org.scottishtecharmy.soundscape.utils.distance
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxCorners
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxOfLineString
-import org.scottishtecharmy.soundscape.utils.getCenterOfBoundingBox
-import org.scottishtecharmy.soundscape.utils.getDestinationCoordinate
-import org.scottishtecharmy.soundscape.utils.getDistanceToFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovIntersectionFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovRoadsFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getIntersectionRoadNames
-import org.scottishtecharmy.soundscape.utils.getIntersectionRoadNamesRelativeDirections
-import org.scottishtecharmy.soundscape.utils.getIntersectionsFOVFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getIntersectionsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getNearestIntersection
-import org.scottishtecharmy.soundscape.utils.getNearestRoad
-import org.scottishtecharmy.soundscape.utils.getQuadrants
-import org.scottishtecharmy.soundscape.utils.getRelativeDirectionsPolygons
-import org.scottishtecharmy.soundscape.utils.getRoadBearingToIntersection
-import org.scottishtecharmy.soundscape.utils.getRoadsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getRoadsFovFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getTileXY
-import org.scottishtecharmy.soundscape.utils.getXYTile
-import org.scottishtecharmy.soundscape.utils.lineStringIsCircular
-import org.scottishtecharmy.soundscape.utils.mergeRoadAndDirectionFeatures
-import org.scottishtecharmy.soundscape.utils.polygonContainsCoordinates
-import org.scottishtecharmy.soundscape.utils.removeDuplicateOsmIds
-import org.scottishtecharmy.soundscape.utils.removeDuplicates
+import org.scottishtecharmy.soundscape.geoengine.utils.RelativeDirections
+import org.scottishtecharmy.soundscape.geoengine.utils.calculateCenterOfCircle
+import org.scottishtecharmy.soundscape.geoengine.utils.createTriangleFOV
+import org.scottishtecharmy.soundscape.geoengine.utils.distance
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxCorners
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxOfLineString
+import org.scottishtecharmy.soundscape.geoengine.utils.getCenterOfBoundingBox
+import org.scottishtecharmy.soundscape.geoengine.utils.getDestinationCoordinate
+import org.scottishtecharmy.soundscape.geoengine.utils.getDistanceToFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovIntersectionFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovRoadsFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getIntersectionRoadNames
+import org.scottishtecharmy.soundscape.geoengine.utils.getIntersectionRoadNamesRelativeDirections
+import org.scottishtecharmy.soundscape.geoengine.utils.getIntersectionsFOVFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getIntersectionsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getNearestIntersection
+import org.scottishtecharmy.soundscape.geoengine.utils.getNearestRoad
+import org.scottishtecharmy.soundscape.geoengine.utils.getQuadrants
+import org.scottishtecharmy.soundscape.geoengine.utils.getRelativeDirectionsPolygons
+import org.scottishtecharmy.soundscape.geoengine.utils.getRoadBearingToIntersection
+import org.scottishtecharmy.soundscape.geoengine.utils.getRoadsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getRoadsFovFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.lineStringIsCircular
+import org.scottishtecharmy.soundscape.geoengine.utils.polygonContainsCoordinates
+import org.scottishtecharmy.soundscape.geoengine.utils.removeDuplicateOsmIds
+import org.scottishtecharmy.soundscape.geoengine.utils.removeDuplicates
 
 class RoundaboutsTest {
     //TODO There are a lot of different types of roundabouts so this might take me a while to work out
