+++ app/src/test/java/org/scottishtecharmy/soundscape/TileUtilsTest.kt
@@ -5,44 +5,43 @@ import org.scottishtecharmy.soundscape.geojsonparser.geojson.GeoMoshi
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Point
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Polygon
-import org.scottishtecharmy.soundscape.utils.distanceToIntersection
-import org.scottishtecharmy.soundscape.utils.getBusStopsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getCombinedDirectionPolygons
-import org.scottishtecharmy.soundscape.utils.getCrossingsFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getEntrancesFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovIntersectionFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovPoiFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovRoadsFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getIntersectionRoadNames
-import org.scottishtecharmy.soundscape.utils.getIntersectionsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getNearestIntersection
-import org.scottishtecharmy.soundscape.utils.getNearestPoi
-import org.scottishtecharmy.soundscape.utils.getNearestRoad
-import org.scottishtecharmy.soundscape.utils.getPathsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getPoiFeatureCollectionBySuperCategory
-import org.scottishtecharmy.soundscape.utils.getPointsOfInterestFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getRoadsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getTilesForRegion
-import org.scottishtecharmy.soundscape.utils.getXYTile
-import org.scottishtecharmy.soundscape.utils.polygonContainsCoordinates
+import org.scottishtecharmy.soundscape.geoengine.utils.distanceToIntersection
+import org.scottishtecharmy.soundscape.geoengine.utils.getBusStopsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getCombinedDirectionPolygons
+import org.scottishtecharmy.soundscape.geoengine.utils.getCrossingsFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getEntrancesFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovIntersectionFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovPoiFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovRoadsFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getIntersectionRoadNames
+import org.scottishtecharmy.soundscape.geoengine.utils.getIntersectionsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getNearestIntersection
+import org.scottishtecharmy.soundscape.geoengine.utils.getNearestPoi
+import org.scottishtecharmy.soundscape.geoengine.utils.getNearestRoad
+import org.scottishtecharmy.soundscape.geoengine.utils.getPathsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getPoiFeatureCollectionBySuperCategory
+import org.scottishtecharmy.soundscape.geoengine.utils.getPointsOfInterestFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getRoadsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getTilesForRegion
+import org.scottishtecharmy.soundscape.geoengine.utils.getXYTile
+import org.scottishtecharmy.soundscape.geoengine.utils.polygonContainsCoordinates
 import com.squareup.moshi.Moshi
 import org.junit.Assert
 import org.junit.Test
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Feature
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LineString
-import org.scottishtecharmy.soundscape.utils.TileGrid.Companion.GRID_SIZE
-import org.scottishtecharmy.soundscape.utils.TileGrid.Companion.getTileGrid
-import org.scottishtecharmy.soundscape.utils.createTriangleFOV
-import org.scottishtecharmy.soundscape.utils.distance
-import org.scottishtecharmy.soundscape.utils.explodeLineString
-import org.scottishtecharmy.soundscape.utils.explodePolygon
-import org.scottishtecharmy.soundscape.utils.getDistanceToFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getGpsFromNormalizedMapCoordinates
-import org.scottishtecharmy.soundscape.utils.getNormalizedFromGpsMapCoordinates
-import org.scottishtecharmy.soundscape.utils.getSuperCategoryElements
-import org.scottishtecharmy.soundscape.utils.removeDuplicateOsmIds
-import org.scottishtecharmy.soundscape.utils.sortedByDistanceTo
-import org.scottishtecharmy.soundscape.utils.traceLineString
+import org.scottishtecharmy.soundscape.geoengine.utils.TileGrid.Companion.GRID_SIZE
+import org.scottishtecharmy.soundscape.geoengine.utils.TileGrid.Companion.getTileGrid
+import org.scottishtecharmy.soundscape.geoengine.utils.createTriangleFOV
+import org.scottishtecharmy.soundscape.geoengine.utils.explodeLineString
+import org.scottishtecharmy.soundscape.geoengine.utils.explodePolygon
+import org.scottishtecharmy.soundscape.geoengine.utils.getDistanceToFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getGpsFromNormalizedMapCoordinates
+import org.scottishtecharmy.soundscape.geoengine.utils.getNormalizedFromGpsMapCoordinates
+import org.scottishtecharmy.soundscape.geoengine.utils.getSuperCategoryElements
+import org.scottishtecharmy.soundscape.geoengine.utils.removeDuplicateOsmIds
+import org.scottishtecharmy.soundscape.geoengine.utils.sortedByDistanceTo
+import org.scottishtecharmy.soundscape.geoengine.utils.traceLineString
 
 class TileUtilsTest {
     private val moshi = GeoMoshi.registerAdapters(Moshi.Builder()).build()
