+++ app/src/test/java/org/scottishtecharmy/soundscape/VisuallyCheckOutput.kt
@@ -3,43 +3,35 @@ package org.scottishtecharmy.soundscape
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Feature
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.FeatureCollection
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.GeoMoshi
-import org.scottishtecharmy.soundscape.geojsonparser.geojson.LineString
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.Point
-import org.scottishtecharmy.soundscape.geojsonparser.geojson.Polygon
-import org.scottishtecharmy.soundscape.utils.Quadrant
-import org.scottishtecharmy.soundscape.utils.RelativeDirections
-import org.scottishtecharmy.soundscape.utils.circleToPolygon
-import org.scottishtecharmy.soundscape.utils.cleanTileGeoJSON
-import org.scottishtecharmy.soundscape.utils.createTriangleFOV
-import org.scottishtecharmy.soundscape.utils.getAheadBehindDirectionPolygons
-import org.scottishtecharmy.soundscape.utils.getBoundingBoxCorners
-import org.scottishtecharmy.soundscape.utils.getCenterOfBoundingBox
-import org.scottishtecharmy.soundscape.utils.getCombinedDirectionPolygons
-import org.scottishtecharmy.soundscape.utils.getDestinationCoordinate
-import org.scottishtecharmy.soundscape.utils.getEntrancesFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovIntersectionFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovPoiFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getFovRoadsFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getIndividualDirectionPolygons
-import org.scottishtecharmy.soundscape.utils.getIntersectionRoadNames
-import org.scottishtecharmy.soundscape.utils.getIntersectionsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getLeftRightDirectionPolygons
-import org.scottishtecharmy.soundscape.utils.getNearestIntersection
-import org.scottishtecharmy.soundscape.utils.getPathsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getPoiFeatureCollectionBySuperCategory
-import org.scottishtecharmy.soundscape.utils.getPointsOfInterestFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getPolygonOfBoundingBox
-import org.scottishtecharmy.soundscape.utils.getQuadrants
-import org.scottishtecharmy.soundscape.utils.getReferenceCoordinate
-import org.scottishtecharmy.soundscape.utils.getRelativeDirectionsPolygons
-import org.scottishtecharmy.soundscape.utils.getRoadsFeatureCollectionFromTileFeatureCollection
-import org.scottishtecharmy.soundscape.utils.getTilesForRegion
-import org.scottishtecharmy.soundscape.utils.getXYTile
-import org.scottishtecharmy.soundscape.utils.polygonContainsCoordinates
-import org.scottishtecharmy.soundscape.utils.tileToBoundingBox
+import org.scottishtecharmy.soundscape.geoengine.utils.RelativeDirections
+import org.scottishtecharmy.soundscape.geoengine.utils.circleToPolygon
+import org.scottishtecharmy.soundscape.geoengine.utils.cleanTileGeoJSON
+import org.scottishtecharmy.soundscape.geoengine.utils.createTriangleFOV
+import org.scottishtecharmy.soundscape.geoengine.utils.getAheadBehindDirectionPolygons
+import org.scottishtecharmy.soundscape.geoengine.utils.getBoundingBoxCorners
+import org.scottishtecharmy.soundscape.geoengine.utils.getCenterOfBoundingBox
+import org.scottishtecharmy.soundscape.geoengine.utils.getCombinedDirectionPolygons
+import org.scottishtecharmy.soundscape.geoengine.utils.getDestinationCoordinate
+import org.scottishtecharmy.soundscape.geoengine.utils.getEntrancesFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovIntersectionFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovPoiFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getFovRoadsFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getIndividualDirectionPolygons
+import org.scottishtecharmy.soundscape.geoengine.utils.getIntersectionsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getLeftRightDirectionPolygons
+import org.scottishtecharmy.soundscape.geoengine.utils.getPathsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getPoiFeatureCollectionBySuperCategory
+import org.scottishtecharmy.soundscape.geoengine.utils.getPointsOfInterestFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getPolygonOfBoundingBox
+import org.scottishtecharmy.soundscape.geoengine.utils.getQuadrants
+import org.scottishtecharmy.soundscape.geoengine.utils.getRelativeDirectionsPolygons
+import org.scottishtecharmy.soundscape.geoengine.utils.getRoadsFeatureCollectionFromTileFeatureCollection
+import org.scottishtecharmy.soundscape.geoengine.utils.getTilesForRegion
+import org.scottishtecharmy.soundscape.geoengine.utils.getXYTile
+import org.scottishtecharmy.soundscape.geoengine.utils.tileToBoundingBox
 import com.squareup.moshi.Moshi
-import org.junit.Assert
 import org.junit.Test
 
 // Functions to output GeoJSON strings that can be put into the very useful Geojson.io
